csgo_gui.py
import ctypes
import imgui
import pymeow
from imgui.integrations import pyglet as pygletIntegration
import pyglet
import threading

# Initialize PyMeow and ImGui
meow = pymeow.PyMeow("csgo.exe")  # Make sure to target the CSGO process

# Initialize pyglet and ImGui
window = pyglet.window.Window(visible=False)
pygletIntegration.configure_pyglet(window)

def draw_gui():
    imgui.new_frame()

    imgui.begin("CSGO Internal GUI")
    imgui.text("Welcome to the internal ImGui interface!")
    imgui.text("This is where you can add features like ESP, Aimbot, etc.")
    
    if imgui.button("ESP"):
        # Implement ESP Toggle here
        print("[+] ESP")

    if imgui.button("AIMBOT"):
        print("[+] AIMBOT")

    imgui.end()
    imgui.render()

    # Render the GUI to the window
    pygletIntegration.render(imgui.get_draw_data())

# Start the GUI in a separate thread
def gui_thread():
    while True:
        draw_gui()
        pyglet.clock.tick()

# Main function
def main():
    # Run the GUI in a background thread
    threading.Thread(target=gui_thread, daemon=True).start()

    # This loop makes sure the CSGO process stays in the background and keeps sending input to the overlay
    while True:
        # This is where you would use PyMeow to manipulate the game, for example:
        # - Read player coordinates
        # - Inject data or manipulate memory
        pass

if __name__ == "__main__":
    main()
